requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==24.3.0
autopep8==2.3.1
blinker==1.9.0
blosc==1.11.2
bm3d==4.0.3
bm4d==4.2.5
boxx==0.10.14
cffi==1.17.1
click==8.1.7
cmake==3.30.3
colorama==0.4.6
colour-demosaicing==0.2.6
colour-science==0.4.6
comm==0.2.2
configargparse==1.7
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
depth-pro==0.1
diffusers==0.30.3
docker-pycreds==0.4.0
drjit==1.0.1
exceptiongroup==1.2.2
executing==2.1.0
exifread==3.0.0
fastjsonschema==2.21.1
filelock==3.14.0
flask==3.0.3
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
gitpython==3.1.43
grpcio==1.66.1
h5py==3.12.1
huggingface-hub==0.25.1
imageio==2.36.0
imagesize==1.4.1
importlib-metadata==8.5.0
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
lazy-loader==0.4
lit==18.1.8
lpips==0.1.4
markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmengine==0.10.5
model-index==0.1.11
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
ninja==1.11.1.1
numpy==1.26.4
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openexr==3.3.3
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pidng==4.0.9
pillow-heif==0.20.0
piq==0.8.0
platformdirs==4.3.6
plotly==5.24.1
process-raw==0.2.1
prompt-toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pygments==2.18.0
pynvml==11.5.3
pyopengl==3.1.7
pyparsing==3.1.4
pyquaternion==0.9.9
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
pytz==2023.4
pywavelets==1.7.0
pyyaml==6.0.2
rawpy==0.23.1
referencing==0.35.1
regex==2024.9.11
requests==2.28.2
retrying==1.3.4
rich==13.4.2
rpds-py==0.22.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==60.2.0
six==1.16.0
smmap==5.0.1
snakeviz==2.2.2
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
tokenizers==0.20.0
tomli==2.0.1
torchgeometry==0.1.2
tornado==6.4.1
tqdm==4.65.2
traitlets==5.14.3
transformers==4.45.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==1.26.20
wandb==0.18.2
wcwidth==0.2.13
werkzeug==3.0.4
widgetsnbextension==4.0.13
xformers==0.0.20
yapf==0.40.2
zipp==3.20.2